215459 - rename the provider class and modified the batchSearch url  AB#215459 AB#215801
